Add WorldRenderTownIconEvent
Allows a plugin to alter the IconMarker that is used to mark a town's
homeblock.

Also pulls the clickToolTip/hoverToolTip that was used in the
WorldRenderTownEvent's MarkerOptions.Builder, so changes made in the
WorldRenderTownEvent automatically carry forward to the IconMarker.

My IDE doesn't have this project set up correctly, so I don't have the
ability to build it or see what would be obvious errors due to missing
methods/typos.